[EC-83] refactor: api 인스턴스 accesstoken 제외 경로 추가
