+++ client/src/api/authApi.js
@@ -1,9 +1,8 @@
-import apiInstance from './apiInstance';
-import authInstance from './authInstance';
+import apiInstance from './instance/apiInstance';
 
 export const authApi = {
   login: async (email, password) => {
-    const response = await authInstance.post(
+    const response = await apiInstance.post(
       '/auth/login',
       {
         email: email,
